Port WM2 branch onto next
This patch ports the changes found in the WM2 branch at this
address:

http://devel.windowmaker.info/cgi-bin/darcsweb.cgi?r=wm2;a=summary

onto the next branch. These changes were started by Dan and Alfredo
to port the 2D drawing in WMaker to Cairo. At the state of this
commit, the some of the stuff in WINGs compiles, but the rest still
needs a lot of work.