Added a simple "draw" program
This program is meant to serve as a frame for quickly testing cairo
code. This should be useful for people like me who are not yet too
familiar with cairo.

This should ultimately not be part of the wmaker tree, but I think it
is useful here while we are getting up to speed.